PYIC-4122: Refactored F2F contract test, added DL F2F scenario
