+++ lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/passportCri/ContractTest.java
@@ -55,14 +55,13 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
-//@Disabled("PACT tests should not be run in build pipelines at this time")
+// @Disabled("PACT tests should not be run in build pipelines at this time")
 @ExtendWith(PactConsumerTestExt.class)
 @ExtendWith(MockitoExtension.class)
 @PactTestFor(providerName = "PassportCriProvider")
 @MockServerConfig(hostInterface = "localhost", port = "1234")
 public class ContractTest {
     private static final String TEST_USER = "test-subject";
-    private static final String TEST_ISSUER = "dummyPassportComponentId";
     private static final String IPV_CORE_CLIENT_ID = "ipv-core";
     private static final String PRIVATE_API_KEY = "dummyApiKey";
     private static final Clock CURRENT_TIME =
@@ -145,78 +144,11 @@ public class ContractTest {
             }
             """;
 
-    private static final String VALID_F2F_VC_BODY = """
-              {
-               "sub": "test-subject",
-               "aud": "dummyF2fComponentId",
-               "nbf": 4070908800,
-               "iss": "dummyF2fComponentId",
-               "vc": {
-                 "type": [
-                   "VerifiableCredential",
-                   "IdentityCheckCredential"
-                 ],
-                 "credentialSubject": {
-                   "name": [
-                     {
-                       "nameParts": [
-                         {
-                           "type": "GivenName",
-                           "value": "Kenneth"
-                         },
-                         {
-                           "type": "FamilyName",
-                           "value": "Decerqueira"
-                         }
-                       ]
-                     }
-                   ],
-                   "birthDate": [
-                     {
-                       "value": "1965-07-08"
-                     }
-                   ],
-                   "socialSecurityRecord": [],
-                   "emailAddress": "dev-platform-testing@digital.cabinet-office.gov.uk",
-                   "passport": [
-                     {
-                       "expiryDate": "2030-01-01",
-                       "documentNumber": "321654987"
-                     }
-                   ]
-                 },
-                 "evidence": [
-                   {
-                     "checkDetails": [
-                       {
-                         "identityCheckPolicy": "published",
-                         "checkMethod": "vcrypt"
-                       },
-                       {
-                         "biometricVerificationProcessLevel": 3,
-                         "checkMethod": "bvr"
-                       }
-                     ],
-                     "validityScore": 2,
-                     "verificationScore": 3,
-                     "strengthScore": 4,
-                     "type": "IdentityCheck",
-                     "txn": "eda339dd-aa83-495c-a4d4-75021e9415f9"
-                   }
-                 ]
-               },
-               "jti": "test-jti"
-             }
-            """;
-
     // If we generate the signature in code it will be different each time, so we need to generate a
     // valid signature (using https://jwt.io works well) and record it here so the PACT file doesn't
     // change each time we run the tests.
     private static final String VALID_VC_SIGNATURE =
             "lDEi47XWnjyZ20fU-vb02BV1MSan68AqLUEQxCZGM_r8i4bG06uzdpXOJZeg-Kdhsf-NtWbqb-xM0P36YGwIeg";
-
-    private static final String VALID_F2F_VC_SIGNATURE =
-            "CAMtOsXoWJiNWG5JPOqRoP8Ry-3hyCRqR1VodFVSbNzsXXTn2xjQXK1J3PIxfc8ZOd9IV-TZC3gZvGty-I9CKw";
 
     @Mock private ConfigService mockConfigService;
     @Mock private JWSSigner mockSigner;
@@ -233,7 +165,7 @@ public class ContractTest {
                 .path("/token")
                 .method("POST")
                 .body(
-                        "client_assertion_type=urn%3Aietf%3Aparams%3Aoauth%3Aclient-assertion-type%3Ajwt-bearer&code=dummyAuthCode&grant_type=authorization_code&redirect_uri=https%3A%2F%2Fidentity.staging.account.gov.uk%2Fcredential-issuer%2Fcallback%3Fid%3DukPassport&client_assertion=eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiJ9.eyJpc3MiOiJpcHYtY29yZSIsInN1YiI6Imlwdi1jb3JlIiwiYXVkIjoiZHVtbXlGMmZDb21wb25lbnRJZCIsImV4cCI6NDA3MDkwOTcwMCwianRpIjoidGVzdC1qdGkifQ.vtEjHaqRX6sPKU0Rn7Gc7uHnWPbAbwPCDsGNsbpOmUNPxjTfgrElvbMQW15pqX4sBO7LnXnpW1Ma8eRPn4Vwqw")
+                        "client_assertion_type=urn%3Aietf%3Aparams%3Aoauth%3Aclient-assertion-type%3Ajwt-bearer&code=dummyAuthCode&grant_type=authorization_code&redirect_uri=https%3A%2F%2Fidentity.staging.account.gov.uk%2Fcredential-issuer%2Fcallback%3Fid%3DukPassport&client_assertion=eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiJ9.eyJpc3MiOiJpcHYtY29yZSIsInN1YiI6Imlwdi1jb3JlIiwiYXVkIjoiZHVtbXlQYXNzcG9ydENvbXBvbmVudElkIiwiZXhwIjo0MDcwOTA5NzAwLCJqdGkiOiJTY25GNGRHWHRoWllYU181azg1T2JFb1NVMDRXLUgzcWFfcDZucHYyWlVZIn0.hXYrKJ_W9YItUbZxu3T63gQgScVoSMqHZ43UPfdB8im8L4d0mZPLC6BlwMJSsfjiAyU1y3c37vm-rV8kZo2uyw")
                 .headers(
                         "x-api-key",
                         PRIVATE_API_KEY,
@@ -249,59 +181,6 @@ public class ContractTest {
                                             body.integerType("expires_in");
                                         })
                                 .build())
-                .toPact();
-    }
-    @Pact(provider = "F2FCriProvider", consumer = "IpvCoreBack")
-    public RequestResponsePact validF2FRequestReturnsValidAccessToken(PactDslWithProvider builder) {
-        return builder.given("dummyAuthCode is a valid authorization code")
-                .given("dummyApiKey is a valid api key")
-                .given("dummyF2fComponentId is the F2F CRI component ID")
-                .given(
-                        "F2F CRI uses CORE_BACK_SIGNING_PRIVATE_KEY_JWK to validate core signatures")
-                .uponReceiving("Valid auth code")
-                .path("/token")
-                .method("POST")
-                .body(
-                        "client_assertion_type=urn%3Aietf%3Aparams%3Aoauth%3Aclient-assertion-type%3Ajwt-bearer&code=dummyAuthCode&grant_type=authorization_code&redirect_uri=https%3A%2F%2Fidentity.staging.account.gov.uk%2Fcredential-issuer%2Fcallback%3Fid%3Df2f&client_assertion=eyJ0eXAiOiJKV1QiLCJhbGciOiJFUzI1NiJ9.eyJpc3MiOiJpcHYtY29yZSIsInN1YiI6Imlwdi1jb3JlIiwiYXVkIjoiZHVtbXlGMmZDb21wb25lbnRJZCIsImV4cCI6NDA3MDkwOTcwMCwianRpIjoidGVzdC1qdGkifQ.vtEjHaqRX6sPKU0Rn7Gc7uHnWPbAbwPCDsGNsbpOmUNPxjTfgrElvbMQW15pqX4sBO7LnXnpW1Ma8eRPn4Vwqw")
-                .headers(
-                        "x-api-key",
-                        PRIVATE_API_KEY,
-                        "Content-Type",
-                        "application/x-www-form-urlencoded; charset=UTF-8")
-                .willRespondWith()
-                .status(200)
-                .body(
-                        newJsonBody(
-                                        (body) -> {
-                                            body.stringType("access_token");
-                                            body.stringValue("token_type", "Bearer");
-                                            body.integerType("expires_in");
-                                        })
-                                .build())
-                .toPact();
-    }
-
-    @Pact(provider = "F2FCriProvider", consumer = "IpvCoreBack")
-    public RequestResponsePact validF2fRequestReturnsIssuedCredential(PactDslWithProvider builder)
-            throws Exception {
-        return builder.given("dummyApiKey is a valid api key")
-                .given("dummyAccessToken is a valid access token")
-                .given("test-subject is a valid subject")
-                .given("dummyF2fComponentId is a valid issuer")
-                .given("VC givenName is Kenneth")
-                .given("VC familyName is Decerqueira")
-                .given("VC birthDate is 1932-02-25")
-                .given("VC passport documentNumber is 824159121")
-                .given("VC passport expiryDate is 2030-01-01")
-                .uponReceiving("Valid POST request")
-                .path("/credential")
-                .method("POST")
-                .headers("x-api-key", PRIVATE_API_KEY, "Authorization", "Bearer dummyAccessToken")
-                .willRespondWith()
-                .status(200)
-                .body(
-                        new PactJwtIgnoreSignatureBodyBuilder(
-                                VALID_VC_HEADER, VALID_F2F_VC_BODY, VALID_F2F_VC_SIGNATURE))
                 .toPact();
     }
 
@@ -327,151 +206,6 @@ public class ContractTest {
                         new PactJwtIgnoreSignatureBodyBuilder(
                                 VALID_VC_HEADER, VALID_VC_BODY, VALID_VC_SIGNATURE))
                 .toPact();
-    }
-
-    @Test
-    @PactTestFor(pactMethod = "validF2FRequestReturnsValidAccessToken")
-    void fetchAccessToken_whenCalledAgainstF2FCri_retrievesAValidAccessToken(
-            MockServer mockServer) throws URISyntaxException, JOSEException, CriApiException {
-        // Arrange
-        var credentialIssuerConfig = getMockF2FCredentialIssuerConfig(mockServer);
-
-        when(mockConfigService.getSsmParameter(ConfigurationVariable.JWT_TTL_SECONDS))
-                .thenReturn("900");
-        when(mockConfigService.getCriConfig(any())).thenReturn(credentialIssuerConfig);
-        when(mockConfigService.getCriPrivateApiKey(any())).thenReturn(PRIVATE_API_KEY);
-
-        // Signature generated by jwt.io by debugging the test and getting the client assertion JWT
-        // generated by the test as mocking out the AWSKMS class inside the real signer would be
-        // painful.
-        when(mockSigner.sign(any(), any()))
-                .thenReturn(
-                        new Base64URL(
-                                "hXYrKJ_W9YItUbZxu3T63gQgScVoSMqHZ43UPfdB8im8L4d0mZPLC6BlwMJSsfjiAyU1y3c37vm-rV8kZo2uyw"));
-        when(mockSigner.supportedJWSAlgorithms()).thenReturn(Set.of(JWSAlgorithm.ES256));
-        when(mockSecureTokenHelper.generate())
-                .thenReturn("ScnF4dGXthZYXS_5k85ObEoSU04W-H3qa_p6npv2ZUY");
-
-        // We need to generate a fixed request, so we set the secure token and expiry to constant
-        // values.
-        var underTest =
-                new CriApiService(
-                        mockConfigService, mockSigner, mockSecureTokenHelper, CURRENT_TIME);
-
-        // Act
-        BearerAccessToken accessToken =
-                underTest.fetchAccessToken(
-                        new CriCallbackRequest(
-                                "dummyAuthCode",
-                                credentialIssuerConfig.getClientId(),
-                                "dummySessionId",
-                                "https://identity.staging.account.gov.uk/credential-issuer/callback?id=f2f",
-                                "dummyState",
-                                null,
-                                null,
-                                "dummyIpAddress",
-                                "dummyFeatureSet"),
-                        new CriOAuthSessionItem(
-                                "dummySessionId",
-                                "dummyOAuthSessionId",
-                                "dummyCriId",
-                                "dummyConnection",
-                                900));
-        // Assert
-        assertThat(accessToken.getType(), is(AccessTokenType.BEARER));
-        assertThat(accessToken.getValue(), notNullValue());
-        assertThat(accessToken.getLifetime(), greaterThan(0L));
-    }
-
-    @Test
-    @PactTestFor(pactMethod = "validF2fRequestReturnsIssuedCredential")
-    void testCallToDummyF2fIssueCredential(MockServer mockServer)
-            throws URISyntaxException, CriApiException {
-        // Arrange
-        var credentialIssuerConfig = getMockF2FCredentialIssuerConfig(mockServer);
-
-        ContraIndicatorConfig ciConfig1 = new ContraIndicatorConfig(null, 4, null, null);
-        ContraIndicatorConfig ciConfig2 = new ContraIndicatorConfig(null, 4, null, null);
-
-        Map<String, ContraIndicatorConfig> ciConfigMap = new HashMap<>();
-        ciConfigMap.put("A02", ciConfig1);
-        ciConfigMap.put("A03", ciConfig2);
-
-        when(mockConfigService.getCriConfig(any())).thenReturn(credentialIssuerConfig);
-        when(mockConfigService.getCriPrivateApiKey(any())).thenReturn(PRIVATE_API_KEY);
-        when(mockConfigService.getContraIndicatorConfigMap()).thenReturn(ciConfigMap);
-
-        var verifiableCredentialJwtValidator =
-                new VerifiableCredentialJwtValidator(
-                        mockConfigService,
-                        ((exactMatchClaims, requiredClaims) ->
-                                new FixedTimeJWTClaimsVerifier<>(
-                                        exactMatchClaims,
-                                        requiredClaims,
-                                        Date.from(CURRENT_TIME.instant()))));
-
-        // We need to generate a fixed request, so we set the secure token and expiry to constant
-        // values.
-        var underTest =
-                new CriApiService(
-                        mockConfigService, mockSigner, mockSecureTokenHelper, CURRENT_TIME);
-
-        // Act
-        var verifiableCredentialResponse =
-                underTest.fetchVerifiableCredential(
-                        new BearerAccessToken("dummyAccessToken"),
-                        new CriCallbackRequest(
-                                "dummyAuthCode",
-                                credentialIssuerConfig.getClientId(),
-                                "dummySessionId",
-                                "https://identity.staging.account.gov.uk/credential-issuer/callback?id=f2f",
-                                "dummyState",
-                                null,
-                                null,
-                                "dummyIpAddress",
-                                "dummyFeatureSet"),
-                        new CriOAuthSessionItem(
-                                "dummySessionId",
-                                "dummyOAuthSessionId",
-                                "dummyCriId",
-                                "dummyConnection",
-                                900));
-
-        verifiableCredentialResponse
-                .getVerifiableCredentials()
-                .forEach(
-                        credential -> {
-                            try {
-                                verifiableCredentialJwtValidator.validate(
-                                        credential, credentialIssuerConfig, TEST_USER);
-
-                                JsonNode credentialSubject =
-                                        objectMapper
-                                                .readTree(credential.getJWTClaimsSet().toString())
-                                                .get("vc")
-                                                .get("credentialSubject");
-
-                                JsonNode nameParts =
-                                        credentialSubject.get("name").get(0).get("nameParts");
-                                JsonNode birthDateNode = credentialSubject.get("birthDate").get(0);
-                                JsonNode passportNode = credentialSubject.get("passport").get(0);
-
-                                assertEquals("GivenName", nameParts.get(0).get("type").asText());
-                                assertEquals("FamilyName", nameParts.get(1).get("type").asText());
-                                assertEquals("Kenneth", nameParts.get(0).get("value").asText());
-                                assertEquals("Decerqueira", nameParts.get(1).get("value").asText());
-
-                                assertEquals("2030-01-01", passportNode.get("expiryDate").asText());
-                                assertEquals(
-                                        "321654987", passportNode.get("documentNumber").asText());
-
-                                assertEquals("1965-07-08", birthDateNode.get("value").asText());
-                            } catch (VerifiableCredentialException
-                                     | ParseException
-                                     | JsonProcessingException e) {
-                                throw new RuntimeException(e);
-                            }
-                        });
     }
 
     @Test
@@ -620,8 +354,8 @@ public class ContractTest {
     }
 
     @NotNull
-    private static CredentialIssuerConfig getMockPassportCredentialIssuerConfig(MockServer mockServer)
-            throws URISyntaxException {
+    private static CredentialIssuerConfig getMockPassportCredentialIssuerConfig(
+            MockServer mockServer) throws URISyntaxException {
         return new CredentialIssuerConfig(
                 new URI("http://localhost:" + mockServer.getPort() + "/token"),
                 new URI("http://localhost:" + mockServer.getPort() + "/credential"),
@@ -635,6 +369,7 @@ public class ContractTest {
                 true,
                 false);
     }
+
     @NotNull
     private static CredentialIssuerConfig getMockF2FCredentialIssuerConfig(MockServer mockServer)
             throws URISyntaxException {
